+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- CSS Handles to the app.
+
+### Fixed
+- Update docs.
+
+### Remove
+- Not implemented props `showPromotionCategory` and `showGiftCategory`.
 
 ## [2.13.2] - 2019-08-29
 ### Changed